Add Update-LabBaseImage function
The function slipstreams windows updates into a given base vhdx. (The ISO image remains in the original state.)
The function was born of necessity as my outdated work laptop isn't uefi enabled (which is required by the Update-LabIsoImage function).
I tested the function with a ws2012r2 and a ws2016 base image and successfully deployed labs afterwards.