Feature/sql azure customizations ( Fixes #171) (#177)
* Added check for SQL role properties

in Azure VM deployment to see if customization is necessary, thus deploying a standard VM

* Type-comparison with -is stopped working for Azure objects

* Type comparison of Azure objects

* Azure: Add-LabIsoImageDefinition for single files

Did not work properly
SQL Setup updated to skip Azure VMs that need no initialization (i.e. will be installed from ISO)

* Azure: Sync-LabAzureLabSource can now use Filter

To be able to sync only single files
New parameter is being used in Add-LabIsoImageDefinition to sync the file if it is missing on Azure

* SQL Azure: Typo fixed

* Azure Iso Mounting: Sync ISO files automatically

* Azure ISO Upload with Add-LabIsoImageDefinition improved

* Fix for Azure SQL installation

Too many DVD drives were returned, so we are filtering on MediaLoaded
Also, the local lab sources will be used to download the SQL files

* SQL on Azure: Calls to Checkpoint and Save-Vm modified

* Azure: Implemented ISO Dismount

Modified Mount-LWAzureIsoImage to copy the ISO in case the connection to the lab sources share is broken

* SQL installation: Removed scheduled task in favor of process

Since unattended setup in Azure did not work. Invoke-LabCommand with CredSSP will also allow setup to work.

* Connect-LabVm: Fixed empty connection string

* Reboot added again since AL waits for it

* Azure SQL: Additional Setup parameter for Azure

Otherwise, SQL Setup runs into AccessDenied while looking for updates

* AutoLogon for Azure machines

Added new cmdlet to enable auto-logon for lab installation user

* Timing autologon

* Added function Test-LabAutoLogon

to test if autologon is enabled and that there is indeed a logon session active on the VM
Fixed issue with SQL sample DB download folder not being created

* Added logging to Test-LabAutoLogon

* Test-LabAutoLogon narrowed down to interactive sessions only

* CIM filter corrected

* Test-LabAutoLogon added EA SilentlyContinue

* Position of Test-LabAutoLogon moved